[Editor] - updated to wxWidgets 2.9 (in progress)

+++ Core/Source/IceGOCView.cpp
@@ -9,7 +9,7 @@ namespace Ice
 
 	GOCEditorVisualised::~GOCEditorVisualised()
 	{
-		if (mEditorVisual)
+		if (mEditorVisual && mOwnerGO)
 		{
 			GetNode()->detachObject(mEditorVisual);
 			Main::Instance().GetOgreSceneMgr()->destroyEntity(mEditorVisual);
